Fix warnings when building with --no-default-features.
